agents/requirements.txt
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.0
langchain==0.3.23
langchain-community==0.3.21
langchain-openai==0.3.12
langgraph==0.3.27
redis==5.2.1
langchain-redis==0.2.0